feat(mcdc): short-circuit clause evaluation with u32 eval-flag encoding
Replace eager pre-evaluation with a short-circuit tree that faithfully models
&&/|| semantics. Each leaf clause sets __d{id}_e[i]=true and __d{id}_c[i]=value
only when reached; masked clauses remain false.

The observation encoding changes from u16 to u32:
  bits 0..N-1:   clause values
  bits N..2N-1:  eval flags (0 = masked by short-circuit)
  bit 2N:        decision outcome

Independence check updated to treat masked clauses as unconstrained — they impose
no requirement on the Unique-Cause pair. This makes the DO-178C criterion
practically achievable without relaxing strictness.

Updates: emit_stmts.rs (sc tree emission), mcdc_instr.rs (u32 encoding +
is_clause_covered with mask handling), main.rs (u32 obs parsing), transpiler
tests, spec 010 Req 3 + Req 5, ADR-0015 Decisions 1-3.

+++ src/main.rs
@@ -653,14 +653,14 @@ fn cmd_mcdc(path: &str, quiet: bool, verbose: bool) {
 
     // Parse observations.
     let raw_obs = fs::read_to_string(&mcdc_out_path).unwrap_or_default();
-    let observations: Vec<Vec<u16>> = raw_obs
+    let observations: Vec<Vec<u32>> = raw_obs
         .lines()
         .map(|line| {
             if line.is_empty() {
                 Vec::new()
             } else {
                 line.split(',')
-                    .filter_map(|hex| u16::from_str_radix(hex.trim(), 16).ok())
+                    .filter_map(|hex| u32::from_str_radix(hex.trim(), 16).ok())
                     .collect()
             }
         })
